# doc/openapi.yaml
openapi: "3.0.3"
info:
  title: AiR ORCHESTRATOR API
  description: |
    REST API сервиса AiR ORCHESTRATOR. Большинство защищённых маршрутов требуют
    заголовок `Authorization: Bearer <token>`. WebSocket-маршруты принимают
    токен через query-параметр `token` или через `Sec-WebSocket-Protocol`.
  version: "1.0.0"
servers:
  - url: https://localhost:8443
    description: Локальная разработка (TLS)
  - url: http://localhost:8080
    description: Production (за nginx)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /nota/mail:
    get:
      tags: [nota]
      summary: Получить email уведомлений
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /nota:
    get:
      tags: [nota]
      summary: Получить настройки уведомлений
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
    post:
      tags: [nota]
      summary: Сохранить настройки уведомлений
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
    delete:
      tags: [nota]
      summary: Удалить канал уведомлений
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /nota/code:
    post:
      tags: [nota]
      summary: Отправить код верификации
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /nota/events:
    post:
      tags: [nota]
      summary: Сохранить события уведомлений
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK

  # ─────────────── NOTIFICATION (outbound) ───────────────
  /notification/mail:
    post:
      tags: [notification]
      summary: Отправить email-уведомление
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /notification/telega:
    post:
      tags: [notification]
      summary: Отправить Telegram-уведомление
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /notification/instant:
    post:
      tags: [notification]
      summary: Отправить мгновенное уведомление
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK

  # ─────────────── USER ───────────────
...
  /operators:
    get:
      tags: [operators]
      summary: Список операторов
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
    post:
      tags: [operators]
      summary: Сохранить операторов
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK

  # ─────────────── SERVICES ───────────────
...